Support Rubocop 0.51, drop JRuby 1.7 support (#5)
* Support Rubocop 0.51, drop JRuby 1.7 support

Newer versions of Rubocop only support Ruby 2.1+. We cannot use these versions with the gem, unless we change the TargetRubyVersion in our .rubocop.yml file.

The simplest approach for fixing this is to drop JRuby 1.7 support. It has reached EOL anyway, and we have little internal need for the gem with JRuby 1.7. Hence, I suggest we drop it; _if_ we ever need it, we can fork out from our latest release and take it from there and onwards.

* gemspec: Update required_ruby_version

* Travis: remove JRuby 9.0.5.0 CI runs

That JRuby version is considered buggy by ruby_dep and its usage is generally discouraged.